Rename `shared/types` to `shared/ontology-types`
