+++ packages/hash/api/src/graph/system-user.ts
@@ -1,7 +1,7 @@
 import { systemUserShortname } from "@hashintel/hash-shared/environment";
 import { Logger } from "@hashintel/hash-backend-utils/logger";
 import { GraphApi } from "@hashintel/hash-graph-client";
-import { types } from "@hashintel/hash-shared/types";
+import { types } from "@hashintel/hash-shared/ontology-types";
 import { getEntities } from "@hashintel/hash-subgraph/src/stdlib/element/entity";
 import {
   extractEntityUuidFromEntityId,
