+++ packages/hash/frontend/src/components/hooks/useOrgs.ts
@@ -1,6 +1,6 @@
 import { useMemo } from "react";
 import { useQuery } from "@apollo/client";
-import { types } from "@hashintel/hash-shared/types";
+import { types } from "@hashintel/hash-shared/ontology-types";
 import { getRoots } from "@hashintel/hash-subgraph/src/stdlib/roots";
 import { Subgraph, SubgraphRootTypes } from "@hashintel/hash-subgraph";
 import {
