+++ packages/hash/frontend/src/components/hooks/useUpdateAuthenticatedUser.ts
@@ -1,7 +1,7 @@
 import { useMutation } from "@apollo/client";
 import { useCallback, useState } from "react";
 
-import { types } from "@hashintel/hash-shared/types";
+import { types } from "@hashintel/hash-shared/ontology-types";
 import { extractBaseUri } from "@blockprotocol/type-system";
 import { GraphQLError } from "graphql";
 import { getRoots } from "@hashintel/hash-subgraph/src/stdlib/roots";
