+++ packages/hash/frontend/src/lib/org.ts
@@ -1,5 +1,5 @@
 import { extractBaseUri } from "@blockprotocol/type-system";
-import { types } from "@hashintel/hash-shared/types";
+import { types } from "@hashintel/hash-shared/ontology-types";
 import {
   Subgraph,
   EntityEditionId,
