+++ packages/hash/frontend/src/pages/[account-slug]/entities/[entity-uuid].page/entity-editor/properties-section/property-table/cells/value-cell/value-cell-editor.tsx
@@ -1,4 +1,4 @@
-import { types } from "@hashintel/hash-shared/types";
+import { types } from "@hashintel/hash-shared/ontology-types";
 import { ValueCellEditorComponent } from "./types";
 import { ArrayEditor } from "./value-cell-editor/array-editor";
 import { BooleanEditor } from "./value-cell-editor/boolean-editor";
