+++ packages/hash/frontend/src/pages/[account-slug]/entities/[entity-uuid].page/entity-editor/properties-section/property-table/cells/value-cell/value-cell-editor/number-or-string-editor.tsx
@@ -1,5 +1,5 @@
 import { TextField } from "@hashintel/hash-design-system";
-import { types } from "@hashintel/hash-shared/types";
+import { types } from "@hashintel/hash-shared/ontology-types";
 import produce from "immer";
 import { ValueCellEditorComponent } from "../types";
