+++ packages/hash/frontend/src/pages/[account-slug]/types/entity-type/[entity-type-id].page/definition-tab/property-list-card/shared/property-type-form.tsx
@@ -7,7 +7,7 @@ import {
   TextField,
 } from "@hashintel/hash-design-system";
 import { frontendUrl } from "@hashintel/hash-shared/environment";
-import { generateBaseTypeId } from "@hashintel/hash-shared/types";
+import { generateBaseTypeId } from "@hashintel/hash-shared/ontology-types";
 import { versionedUriFromComponents } from "@hashintel/hash-subgraph/src/shared/type-system-patch";
 import { getPropertyTypeById } from "@hashintel/hash-subgraph/src/stdlib/element/property-type";
 import {
