+++ packages/hash/frontend/src/pages/[account-slug]/types/entity-type/[entity-type-id].page/definition-tab/property-list-card/shared/property-type-form/data-type-selector/shared/expected-values-options.ts
@@ -1,6 +1,6 @@
 import { faList, faListCheck } from "@fortawesome/free-solid-svg-icons";
 import { theme } from "@hashintel/hash-design-system";
-import { types } from "@hashintel/hash-shared/types";
+import { types } from "@hashintel/hash-shared/ontology-types";
 import { fa100 } from "../../../../../../../../../../../shared/icons/pro/fa-100";
 import { faCube } from "../../../../../../../../../../../shared/icons/pro/fa-cube";
 import { faCubes } from "../../../../../../../../../../../shared/icons/pro/fa-cubes";
