+++ packages/hash/frontend/src/pages/[account-slug]/types/entity-type/[entity-type-id].page/definition-tab/property-list-card/shared/property-type-form/shared/data-type-options.ts
@@ -1,4 +1,4 @@
-import { types } from "@hashintel/hash-shared/types";
+import { types } from "@hashintel/hash-shared/ontology-types";
 
 export const dataTypeOptions = [
   types.dataType.text.dataTypeId,
