+++ packages/hash/frontend/src/pages/entity-editor.page.tsx
@@ -3,7 +3,7 @@ import { useRouter } from "next/router";
 import { Container, Typography } from "@mui/material";
 import { ValueOrArray, Array } from "@blockprotocol/type-system";
 import { Button } from "@hashintel/hash-design-system";
-import { types } from "@hashintel/hash-shared/types";
+import { types } from "@hashintel/hash-shared/ontology-types";
 import { Entity, Subgraph, SubgraphRootTypes } from "@hashintel/hash-subgraph";
 import { getRoots } from "@hashintel/hash-subgraph/src/stdlib/roots";
 import { getEntityTypeById } from "@hashintel/hash-subgraph/src/stdlib/element/entity-type";
