+++ packages/hash/frontend/src/pages/new/types/entity-type.page.tsx
@@ -1,7 +1,7 @@
 import { EntityType } from "@blockprotocol/type-system";
 import { Button, TextField } from "@hashintel/hash-design-system";
 import { frontendUrl } from "@hashintel/hash-shared/environment";
-import { generateBaseTypeId } from "@hashintel/hash-shared/types";
+import { generateBaseTypeId } from "@hashintel/hash-shared/ontology-types";
 import {
   Box,
   Container,
