+++ packages/hash/shared/src/ontology-types.ts
